Login y registro para backend listo

+++ routes/web.php
@@ -2,11 +2,12 @@
 
 use App\Http\Controllers\admin\CandyController as AdminCandyController;
 use App\Http\Controllers\CandyController;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 
 
 Route::get('/', function () {
-    return view('welcome');
+    return redirect()->route('web.candies.index');
 });
 
 Route::get('candies', [CandyController::class, 'index']) ->name('web.candies.index');
@@ -17,9 +18,14 @@ Route::post('/cart/remove/{id}', [CandyController::class, 'remove'])->name('web.
 Route::post('/cart/buy', [CandyController::class, 'buy'])->name('web.candies.buy');
 Route::get('/cart', [CandyController::class, 'cart'])->name('web.candies.cart');
 
-Route::get('admin/candies', [AdminCandyController::class, 'index']) ->name('admin.index');
-Route::get('admin/candies/create', [AdminCandyController::class, 'create']) ->name('admin.create');
-Route::post('admin/candies', [AdminCandyController::class, 'store']) ->name('admin.store');
-Route::get('admin/candies/{id}/edit', [AdminCandyController::class, 'edit']) ->name('admin.edit');
-Route::post('admin/candies/{id}', [AdminCandyController::class, 'update']) ->name('admin.update');
-Route::delete('admin/candies/{id}', [AdminCandyController::class, 'destroy'])->name('admin.destroy');
+Route::get('admin/candies', [AdminCandyController::class, 'index']) ->name('admin.index') ->middleware('auth');
+Route::get('admin/candies/create', [AdminCandyController::class, 'create']) ->name('admin.create') ->middleware('auth');
+Route::post('admin/candies', [AdminCandyController::class, 'store']) ->name('admin.store') ->middleware('auth');
+Route::get('admin/candies/{id}/edit', [AdminCandyController::class, 'edit']) ->name('admin.edit') ->middleware('auth');
+Route::post('admin/candies/{id}', [AdminCandyController::class, 'update']) ->name('admin.update') ->middleware('auth');
+Route::delete('admin/candies/{id}', [AdminCandyController::class, 'destroy'])->name('admin.destroy') ->middleware('auth');
+
+
+Auth::routes();
+
+Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
